+++ core/src/commonMain/kotlin/com/jsontextfield/departurescreen/core/ui/menu/Action.kt
@@ -1,4 +1,4 @@
-package com.jsontextfield.departurescreen.ui.menu
+package com.jsontextfield.departurescreen.core.ui.menu
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.vector.ImageVector
