+++ core/src/commonMain/kotlin/com/jsontextfield/departurescreen/core/ui/menu/OverflowMenu.kt
@@ -1,8 +1,8 @@
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.runtime.Composable
-import com.jsontextfield.departurescreen.ui.menu.Action
-import com.jsontextfield.departurescreen.ui.menu.NestedOverflowMenuItem
-import com.jsontextfield.departurescreen.ui.menu.OverflowMenuItem
+import com.jsontextfield.departurescreen.core.ui.menu.Action
+import com.jsontextfield.departurescreen.core.ui.menu.NestedOverflowMenuItem
+import com.jsontextfield.departurescreen.core.ui.menu.OverflowMenuItem
 
 @Composable
 fun OverflowMenu(
