+++ core/src/commonMain/kotlin/com/jsontextfield/departurescreen/core/ui/menu/OverflowMenuItem.kt
@@ -1,4 +1,4 @@
-package com.jsontextfield.departurescreen.ui.menu
+package com.jsontextfield.departurescreen.core.ui.menu
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Check
