+++ core/src/commonMain/kotlin/com/jsontextfield/departurescreen/core/ui/menu/PopupMenu.kt
@@ -1,4 +1,4 @@
-package com.jsontextfield.departurescreen.ui.menu
+package com.jsontextfield.departurescreen.core.ui.menu
 
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.runtime.Composable
